synthetic excell data generator
A code that converts customer name, ID, customer number, city of residence and credit limit to excell file. This code is saved on the desktop in xls format. Places to customise and change in the code were written.pandas, openpyxl and random packages must be down for the code to work. Made in python version 3.13